Move user typescript files into $lib/user
