+++ apps/desktop/src/components/AuthorizationBanner.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { UserService } from '$lib/stores/user';
+	import { UserService } from '$lib/user/userService';
 	import { getContext } from '@gitbutler/shared/context';
 	import Button from '@gitbutler/ui/Button.svelte';
 	import Icon from '@gitbutler/ui/Icon.svelte';
