+++ apps/desktop/src/components/BranchLaneContextMenu.svelte
@@ -4,7 +4,7 @@
 	import ContextMenuSection from '$components/ContextMenuSection.svelte';
 	import { getForgePrService } from '$lib/forge/interface/forgePrService';
 	import { updatePrDescriptionTables } from '$lib/forge/shared/prFooter';
-	import { User } from '$lib/stores/user';
+	import { User } from '$lib/user/user';
 	import { BranchController } from '$lib/vbranches/branchController';
 	import { BranchStack } from '$lib/vbranches/types';
 	import { getContext, getContextStore } from '@gitbutler/shared/context';
