+++ apps/desktop/src/components/CloudProjectSettings.svelte
@@ -3,7 +3,7 @@
 	import ProjectConnectModal from '$components/ProjectConnectModal.svelte';
 	import Section from '$components/Section.svelte';
 	import { ProjectsService, ProjectService } from '$lib/backend/projects';
-	import { User } from '$lib/stores/user';
+	import { User } from '$lib/user/user';
 	import * as toasts from '$lib/utils/toasts';
 	import { getContext, getContextStore } from '@gitbutler/shared/context';
 	import RegisterInterest from '@gitbutler/shared/interest/RegisterInterest.svelte';
