+++ apps/desktop/src/components/DecorativeSplitView.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import AccountLink from '$components/AccountLink.svelte';
 	import gbLogoSvg from '$lib/assets/gb-logo.svg?raw';
-	import { User } from '$lib/stores/user';
+	import { User } from '$lib/user/user';
 	import { openExternalUrl } from '$lib/utils/url';
 	import { getContextStore } from '@gitbutler/shared/context';
 	import Icon from '@gitbutler/ui/Icon.svelte';
