+++ apps/desktop/src/components/ProjectSetupTarget.svelte
@@ -8,7 +8,7 @@
 	import { ProjectsService, Project } from '$lib/backend/projects';
 	import { projectAiGenEnabled } from '$lib/config/config';
 	import { platformName } from '$lib/platform/platform';
-	import { UserService } from '$lib/stores/user';
+	import { UserService } from '$lib/user/userService';
 	import { unique } from '$lib/utils/array';
 	import { getBestBranch, getBestRemote, getBranchRemoteFromRef } from '$lib/utils/branch';
 	import { getContext } from '@gitbutler/shared/context';
