+++ apps/desktop/src/components/WelcomeSigninAction.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import WelcomeAction from './WelcomeAction.svelte';
 	import signinSvg from '$lib/assets/signin.svg?raw';
-	import { UserService } from '$lib/stores/user';
+	import { UserService } from '$lib/user/userService';
 	import { getContext } from '@gitbutler/shared/context';
 	import LinkButton from '@gitbutler/ui/LinkButton.svelte';
