+++ apps/desktop/src/lib/analytics/sentry.ts
@@ -1,6 +1,6 @@
 import * as Sentry from '@sentry/sveltekit';
 import { type Span } from '@sentry/sveltekit';
-import type { User } from '$lib/stores/user';
+import type { User } from '$lib/user/user';
 import { dev } from '$app/environment';
 import { PUBLIC_SENTRY_ENVIRONMENT } from '$env/static/public';
