+++ apps/desktop/src/lib/user/user.ts
@@ -1,0 +1,16 @@
+export class User {
+	id!: number;
+	name: string | undefined;
+	given_name: string | undefined;
+	family_name: string | undefined;
+	email!: string | undefined;
+	picture!: string;
+	locale!: string | undefined;
+	created_at!: string;
+	updated_at!: string;
+	access_token!: string;
+	role: string | undefined;
+	supporter!: boolean;
+	github_access_token: string | undefined;
+	github_username: string | undefined;
+}
