+++ apps/desktop/src/lib/user/userService.ts
@@ -1,3 +1,4 @@
+import { User } from './user';
 import { resetSentry, setSentryUser } from '$lib/analytics/sentry';
 import { invoke } from '$lib/backend/ipc';
 import { showError } from '$lib/notifications/toasts';
@@ -144,21 +145,4 @@ export class UserService {
 			headers: { 'Content-Type': undefined }
 		});
 	}
-}
-
-export class User {
-	id!: number;
-	name: string | undefined;
-	given_name: string | undefined;
-	family_name: string | undefined;
-	email!: string | undefined;
-	picture!: string;
-	locale!: string | undefined;
-	created_at!: string;
-	updated_at!: string;
-	access_token!: string;
-	role: string | undefined;
-	supporter!: boolean;
-	github_access_token: string | undefined;
-	github_username: string | undefined;
 }
